backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ML and image processing
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# PDF processing
PyMuPDF==1.23.8
textstat==0.7.3
nltk==3.8.1

# Archive extraction
rarfile==4.1
py7zr==0.20.8

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2